+++ CHANGELOG.md
@@ -4,6 +4,7 @@
 - Fixed typo where antiraid DM's had incorrect grammar for the punishment.
 - Fixed bug where, in some circumstances, antiraid log messages would be duplicated.
 - Fixed a bug where users DM'd with an antiraid DM notification would not be called by their proper name.
+- The ability for deleted message's contents to be archived has now been converted to a premium-only feature.
 
 ### Version 3.1.1
 - Fixed bug with antiraid where users would not be actioned or logged properly.